Move silx.gui.plot.matplotlib to silx.gui.utils.matplotlib
